requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
tabulate>=0.9.0
wcwidth>=0.2.0
zai-sdk
